// admin-frontend/src/api/generated/model/roleAssignablePermissionVO.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Admin Backend API
 * 后台管理系统后端 API 文档，前端可通过 OpenAPI Schema 自动生成接口调用代码
 * OpenAPI spec version: 1.0.0
 */
import type { AssignableGroupVO } from './assignableGroupVO';

/**
 * 角色可分配权限
 */
export interface RoleAssignablePermissionVO {
  /** 角色ID */
  roleId?: number;
  /** 分组列表 */
  groups?: AssignableGroupVO[];
}