// admin-frontend/src/api/generated/model/assignableGroupVO.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Admin Backend API
 * 后台管理系统后端 API 文档，前端可通过 OpenAPI Schema 自动生成接口调用代码
 * OpenAPI spec version: 1.0.0
 */
import type { AssignableItemVO } from './assignableItemVO';

/**
 * 分组可分配权限
 */
export interface AssignableGroupVO {
  /** 分组标识 */
  groupKey?: string;
  /** 分组名称 */
  groupName?: string;
  groupPermission?: AssignableItemVO;
  /** 未分配的子权限 */
  unassignedChildren?: AssignableItemVO[];
}